+++ batcher/Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 members = [
     "aligned-batcher",
-    "aligned-batcher-client",
-    "aligned-batcher-lib"
+    "aligned-batcher-lib",
+    "aligned"
 ]
 resolver = "2"
