+++ batcher/aligned-batcher-lib/Cargo.toml
@@ -4,15 +4,15 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-alloy-primitives = "0.7.4"
+alloy-primitives = { version = "0.7.4", features = ["serde"] }
 anyhow = "1.0.83"
 lazy_static = "1.4.0"
 log = "0.4.21"
 serde = { version = "1.0.201", features = ["derive"] }
-sp1-sdk = { git = "https://github.com/succinctlabs/sp1.git", rev = "v1.0.0-testnet" }
-bincode = "1.3.3"
-hex = "0.4.3"
+sp1-sdk = { git = "https://github.com/succinctlabs/sp1.git", rev = "v1.0.5-testnet" }
 halo2curves = { version = "0.6.0", default-features = false }
-halo2_backend = { git = "https://github.com/yetanotherco/yet-another-halo2-fork.git", branch = "feat/serde_constraint_system"}
 halo2_proofs = { git = "https://github.com/yetanotherco/yet-another-halo2-fork.git", branch = "feat/serde_constraint_system" }
-ff = "0.13"
+sha3 = "0.10.8"
+lambdaworks-crypto = { version = "0.7.0", features = ["serde"] }
+hex = "0.4.3"
+bincode = "1.3.3"
