+++ batcher/aligned-batcher-lib/src/halo2/mod.rs
@@ -1,1 +1,2 @@
 pub mod ipa;
+pub mod kzg;
