+++ batcher/aligned-batcher/src/types/batch_queue.rs
@@ -4,7 +4,7 @@ use futures_util::stream::SplitSink;
 use tokio::{net::TcpStream, sync::RwLock};
 use tokio_tungstenite::{tungstenite::Message, WebSocketStream};
 
-use super::{VerificationData, VerificationDataCommitment};
+use aligned_batcher_lib::types::{VerificationData, VerificationDataCommitment};
 
 pub(crate) type BatchQueueEntry = (
     VerificationData,
