+++ batcher/aligned-batcher/src/types/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod batch_queue;
