+++ batcher/aligned/generate_proof_and_send.sh
