+++ batcher/aligned/src/errors.rs
