+++ batcher/aligned/test_files/groth16_bn254/plonk.proof
