+++ batcher/aligned/test_files/groth16_bn254/plonk_pub_input.pub
