+++ batcher/aligned/test_files/halo2_kzg/pub_input.bin
