+++ batcher/aligned/test_files/plonk_bls12_381/plonk.vk
