--- batcher/src/halo2/mod.rs
@@ -1,2 +1,0 @@
-pub mod ipa;
-pub mod kzg;
